Update runtime, fix config, expand error messages
Just got this up and running for my project, so PRing the changes I had
to make:

- The `nodejs` Lambda runtime is no longer supported so I've updated
  [`node-lambda`][1] to 0.11.3 which now defaults to the runtime which
  replaces it, `nodejs6.10`.
- In discovering the above issue as well as some local issues I added
  details to error messages to help me track down causes, so I've left
  these in in the hopes they'll help others.
- Fixed a couple of typos.

I also had trouble understanding what the trigger for the Lambda
function should be. I can see that the function expects a certain
payload format but I failed to find out where that payload is expected
to come from. I ended up making some additional modifications to my own
copy to make this work with a payload from a CodePipeline invocation
(not included in this PR).

Thanks for putting this repository up, it's been a great help!

[1]: https://github.com/motdotla/node-lambda